Bai2\Parsers: simplify exceptions, revert attempt...
...to intercept at higher layers and rethrow with more detail. We may
well add more context-specific details to these error catories later on
where we use these generalized parsers to parse specific record types,
but at this stage I could not discern a benefit to intercepting and
re-throwing; the level of detail is not significantly different between
the three parts of the generalized parsing system, and end users are no
expected to use these directly in general. Right now we encounter errors
exactly where they're discovered and throw a specific message already
fit for purpose.

It's the components that use these parsers (yet to be fleshed out) to
construct the object graph from specific interpretations of incoming
data where it may make sense to attempt to re-throw with
context-specific detail.