Allow dashes in "Customer Reference Number"
We were previously using a relatively strict interpretation of the
specification which did not permit dashes to be embedded within a
Transaction (type-16) record's Customer Reference Number field. This was
causing problems with data encountered in the real world. This commit
makes our parser more permissive when it comes dashes (as well as
underscores) within this field.

Fixes #8. Closes #9.

Automated Test Results:

OK (1578 tests, 2674 assertions)