TransactionParser: strictly parse Customer Reference Number as alpha-numeric (#15)
Previously we tried to loosen the strictness of our parsing of the
Customer Reference Number to suit a specific bank. This was a bad idea.
This PR reverts those changes and restores our parsing of the Customer
Reference Number for transaction records to be based on the spec, which
indicates it should be alpha-numeric (which definitely does not include
things like spaces, ampersands, octothorpes, dashes, and such).

Closes #15.

Automated Test Results:

OK (1573 tests, 2673 assertions)